Avoid race condition when disposing of app
When disposing of the VirtViewerApp, we free the hash table
containing the windows. This causes each window to be freed,
which in turn causes the visibility callback to be invoked.
This can then get NULL pointers from the self->priv->windows
usage.

Blank out priv->windows before unrefing the hashs and add
a check to ensure priv->windows is non-NULL.