remote-viewer: set auto-conf before fullscreen
If a monitor is already in fullscreen, setting auto-conf to true will
not move it until it is re-fullscreen

This was unnoticed, because usually, the first client window is opened
on the first monitor. Also we may argue than relying on g_object_set()
property order is lame and fragile, we better split it in two seperate
calls as this might break upstream.

https://bugzilla.redhat.com/show_bug.cgi?id=872288