kiosk: explicit resize window to fullscreen size
Allow to run the client in kiosk mode with window-manager-less
environment.

This was a conditional workaroud on win32. I am making it
non-conditional to make fullscreen work on non-wm environment. Hence
I don't see the need to refer explicitely to the bug workaround, since
it is no longer something that should be removed, even when bgo 652049
is fixed.