Enable the display before showing the window
This ensures that the display is enabled when it gets its first Allocate event
(which causes a display reconfiguration).  If the display is not enabled at this
point, it won't send down a new monitors_config message until the second
allocation, which may result in the display being disabled until a window is
resized.