Fix rebuild of accelerators menu when loading from file
It's not enough to set the property to notify of its change. Add a
virt_viewer_app_set_enable_accel() helper, and call it after the changes
to accelerators are made when loading from file.

I verified the menu is correctly built when connection from controller
or command line too.